debian: Change how shlibs work
Instead of trying to guess which syslog-ng-core version to depend on,
track the ABI internally, and use a virtual package for modules to
depend on. The virtual package (`syslog-ng-abi-3.6-0` at the moment)
will track the SONAME of libsyslog-ng.so.

This way, we can have the loosest possible dependencies for third-party
and Incubator modules, meaning they do not need to get recompiled for
every single syslog-ng release, nor do we need to guess versions and ABI
stability.

As such, these changes resolve algernon/syslog-ng-debian#1.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>