debian: Workaround for Makefile rebuilding
Since upstream uses a monolithic Makefile, debian/Makefile.am is
included in it, which means that if debian/Makefile.am has a newer date
than Makefile.am, or if debian/Makefile.in does not exist, Makefile.in
will be regenerated.

That is something we want to avoid when building from a dist tarball. To
fix this, when exporting the sources, after doing an overlay, we'll not
only touch debian/Makefile.am, but debian/Makefile.in too, and use the
top-level Makefile.* as references.

Reported-by: Viktor Tusa <viktor.tusa@balabit.com>
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>