Improved python support
This depends on balabit/syslog-ng#686, but to be able to progress
further, here it is. We add dh-python to build-depends, replace
python-minimal with ${python:Depends} in syslog-ng-mod-python's
dependency list, and patch up debian/rules to call dh_python2, and do so
with DH_OPTIONS cleared (because dh_python2 barfs on --dbg-package).

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>