Try again with tagged release
Now, after some time I'm a bit more confident that using a naked git release
is better than the full-featured and all-including release. First, it is
much smaller. Especially because it does not contain all of the third party
libraries which are not used in Debian. Second reason is that this can be
followed much easily, because this is the same as any particular state in
the git repo, so a snapshot can be more easily used if this is the daily
routine.